backend/requirements.txt
# Cass Vessel Backend Dependencies
# Core API
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Claude API - Both raw and Agent SDK
anthropic>=0.39.0
claude-agent-sdk>=0.1.0

# OpenAI API (optional - for multi-LLM support)
openai>=1.0.0

# Vector Database for Memory
chromadb>=0.4.22
numpy<2.0  # Compatibility fix

# Async support
anyio>=4.0.0

# Audio/Speech - Piper for local neural TTS
piper-tts>=1.3.0
pydub>=0.25.0  # For WAV to MP3 conversion
# edge-tts>=6.1.0  # Currently broken (401 auth errors)
# elevenlabs==0.2.27  # For later if we want voice cloning

# Utilities
pydantic>=2.5.3
websockets>=12.0
python-multipart>=0.0.6

# For watch/sensor integration (later)
# bleak==0.21.1  # Bluetooth LE
